blacked, code cleanup
